Consolidate local memory override unit tests
